Fix new pylint warnings.
They are due to the fact that types for arrow are now present. We now make
the int -> Decimal conversion in parse_datetime_duration() explicitly.

+++ src/isoduration/parser/parsing.py
@@ -1,7 +1,7 @@
 from collections import OrderedDict
 from decimal import Decimal, InvalidOperation
 
-import arrow  # type: ignore
+import arrow
 
 from isoduration.parser.exceptions import (
     IncorrectDesignator,
@@ -28,14 +28,14 @@ def parse_datetime_duration(duration_str: str, sign: int) -> Duration:
 
     return Duration(
         DateDuration(
-            years=sign * duration.year,
-            months=sign * duration.month,
-            days=sign * duration.day,
+            years=Decimal(sign * duration.year),
+            months=Decimal(sign * duration.month),
+            days=Decimal(sign * duration.day),
         ),
         TimeDuration(
-            hours=sign * duration.hour,
-            minutes=sign * duration.minute,
-            seconds=sign * duration.second,
+            hours=Decimal(sign * duration.hour),
+            minutes=Decimal(sign * duration.minute),
+            seconds=Decimal(sign * duration.second),
         ),
     )
